Add a script for adding release year to video game descriptions.
Lots of games have just the description "video game". This isn't very
useful, and makes accurately matching items with
external IDs more difficult.

So this fixes that by setting the description to a string
like "2005 video game" for video game items with a publication
date that are currently just using the description "video game".