Create import script for microtransaction.zone.
This imports the microtransaction.zone IDs based on the GiantBomb IDs
already present in Wikidata, since the IDs for MT Zone are
based on GiantBomb.

Includes a check to ensure that the name of the game on MT Zone
matches the name of the item in Wikidata, as a fail-safe
to prevent the introduction of bad data. Also checks to ensure that
the Wikidata item doesn't already have a MT Zone ID before adding
the ID to the item.